core: move enc utils to core
