+++ bitcoins/src/nets.rs
@@ -28,12 +28,14 @@
 //! ```
 use std::marker::PhantomData;
 
-use coins_core::{enc::AddressEncoder, nets::Network};
+use coins_core::{
+    enc::{AddressEncoder, EncodingError},
+    nets::Network
+};
 
 use crate::{
     builder::BitcoinTxBuilder,
     enc::{
-        bases::EncodingError,
         encoder::{Address, BitcoinEncoderMarker, MainnetEncoder, SignetEncoder, TestnetEncoder},
     },
     types::{
