+++ core/Cargo.toml
@@ -8,6 +8,8 @@ repository = "https://github.com/summa-tx/bitcoins-rs"
 license = "LGPL-3.0-or-later"
 
 [dependencies]
+base58check = "0.1.0"
+bech32 = "0.7.2"
 blake2-rfc = "0.2.18"
 bitcoin-spv = "2.0.0"
 hex = "0.4.2"
