+++ core/src/enc/mod.rs
@@ -1,0 +1,8 @@
+//! Holds generalized encoding tools. This includes an `AddressEncoder` and bech32 encoding
+//! and decoding functionality.
+
+pub mod address;
+pub mod bases;
+
+pub use address::*;
+pub use bases::*;
